// src/app/components/citizen/tax/property/no-due-certificate/Models/no-due-certificate.model.ts
export class SearchModel {
    level1Id: number;
    level2Id: number;
    level3Id: number;
    level4Id: number;
    propertyNo: string;
    censusNo: string;
    occupierName: string;
    ownerName: string;
    address: string;
    mobileNo: string;
    pageNo: number;
	pageSize: number;
}


export class ServiceChargeDetail {
    chargeAmount: number;
    chargeName: string;
    chargeNameLocal: string;
    serviceRateMasterId: number
}
export class ServiceCharge {
    serviceName: string;
    serviceNameLocal: string;
    totalAmount: number;
    serviceChargeDetail: ServiceChargeDetail;
    noofCopies: number;
    propertyBasicId: number;
    totalAmountOriginal: number;
    occupierId: number;
    outstandingTax: number;
}

export class PaymentModel {
    occupierId: number;
    outstandingTax: number;
    numberOfCopies: number;
    propertyBasicId: number;
    paymentDetail: PaymentDetail;
    propertyServiceApplicationId: number;
}

export class PaymentDetail {
    accountNumber: string;
    amount: number;
    bankId: number;
    branchId: number;
    instrumentDate: Date;
    instrumentNo: string;
    payeeName: string;
    paymentModeLookupId: number;
}

export class OutstandingDetailModel {
    propertyBasicId: number;
    propertyNo: string;
    ownerName: string;
    mobileNo: string;
    propertyAddress: string;
    outstandingAmount: number;
    occupierOutstandingDetails: OccupierOutstandingDetails;
}

export class OccupierOutstandingDetails {
    occupierCode: number;
    occupierId: number;
    occupierName: string;
    outstandingAmount: number;
    taxRateWiseOutstandingDetails: TaxRateWiseOutstandingDetails;
}
export class TaxRateWiseOutstandingDetails {
    taxRateId: number;
    name: string;
    nameLocal: string;
    outstandingAmount: number
}